account: move 4 forms to Views/
Per #4331:
- Components::AccountProfileForm
    -> Views::Controllers::Account::Profile::Form
- Components::AccountSignupForm
    -> Views::Controllers::Account::Form
    (keeps explicit id 'account_signup_form' — integration test
    uses it in within(...) blocks; the default Views/-derived id
    would be 'account_form' which loses the 'signup' specificity)
- Components::LoginForm
    -> Views::Controllers::Account::Login::Form
- Components::EmailNewPasswordForm
    -> Views::Controllers::Account::Login::EmailNewPasswordForm
    (keeps descriptive suffix — sibling to Login::Form since the
    login controller has multiple form-bearing pages)

4 ERB callers updated. 56 view + controller + integration tests
pass. Tightened a few forbidden `body, text:` assertions in the
moved test files (per .claude/rules/testing.md).
