+++ app/views/controllers/account/login/email_new_password.html.erb
@@ -6,7 +6,7 @@ add_page_title(:email_new_password_title.t)
   <%= :email_new_password_help.tp + :email_spam_notice.tp %>
 </div>
 
-<%= render(Components::EmailNewPasswordForm.new(
+<%= render(Views::Controllers::Account::Login::EmailNewPasswordForm.new(
       @new_user || User.new,
       action: account_new_password_request_path,
       id: "account_email_new_password_form"
