+++ app/views/controllers/account/profile/edit.html.erb
@@ -6,7 +6,7 @@ add_context_nav(account_profile_edit_tabs)
 
 <div class="row">
   <div class="col-sm-7">
-    <%= render(Components::AccountProfileForm.new(
+    <%= render(Views::Controllers::Account::Profile::Form.new(
       @user,
       copyright_holder: @copyright_holder,
       copyright_year: @copyright_year,
